Finalizado task #19 , no he hecho la opcional que he puesto.

+++ Bets21/src/main/java/domain/Question.java
@@ -57,7 +57,7 @@ public class Question implements Serializable {
 		return false;
 	}
 	
-	public Forecast addForecast(String description, double gain, Question question) {
+	public Forecast addForecast(String description, float gain, Question question) {
 		Forecast f = new Forecast(description, gain, question);
 		forecasts.add(f);
 		return f;
